backend/requirements.txt
# FastAPI and server
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
starlette==0.37.2

# Database
sqlalchemy==2.0.36
alembic==1.13.2

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1

# Configuration
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic-core==2.23.4

# LangGraph and LangChain
langgraph==0.2.40
langchain==0.3.7
langchain-core==0.3.11
langchain-mistralai==0.1.0

# Mistral AI
mistralai==1.2.0

# MCP (Model Context Protocol)
mcp==1.0.0

# HTTP Client
httpx==0.27.0
httpcore==1.0.5

# Utilities
python-dateutil==2.9.0
typing-extensions==4.12.2
annotated-types==0.7.0
anyio==4.4.0
sniffio==1.3.1
idna==3.10
